+++ src/oci/data_flow/models/create_private_endpoint_details.py
@@ -10,7 +10,7 @@ from oci.decorators import init_model_state_from_kwargs
 @init_model_state_from_kwargs
 class CreatePrivateEndpointDetails(object):
     """
-    The create private endpoint details.
+    The details required to create a private endpoint.
     """
 
     def __init__(self, **kwargs):