+++ src/oci/data_flow/models/work_request_summary.py
@@ -10,7 +10,7 @@ from oci.decorators import init_model_state_from_kwargs
 @init_model_state_from_kwargs
 class WorkRequestSummary(object):
     """
-    A Data Flow work request object.
+    A Data Flow work request summary object.
     """
 
     #: A constant which can be used with the operation property of a WorkRequestSummary.